Separation into 2 classes
