+++ ast-visual-studio-extension/ast-visual-studio-extension.csproj
@@ -79,6 +79,7 @@
     <Compile Include="CxExtension\Panels\ResultsTreePanel.cs" />
     <Compile Include="CxExtension\Panels\ResultVulnerabilitiesPanel.cs" />
     <Compile Include="CxExtension\Services\ASCAService.cs" />
+    <Compile Include="CxExtension\Services\ASCAUIManager.cs" />
     <Compile Include="CxExtension\Toolbar\BranchesCombobox.cs" />
     <Compile Include="CxExtension\Toolbar\ComboboxBase.cs" />
     <Compile Include="CxExtension\Toolbar\ProjectsCombobox.cs" />
